Fix package declarations in tests
Gradle (and Maven, its predecessor) group code into folders that distinguish:
- whether it is production code (`main/`) or test code (`test/`)
- which language it is written in - e.g. `java/`, `scala/`, etc...

However, `main/java/` and `test/java/` are not part of the packge.  Rather, the
Java compiler is configured (by Gradle) to look for code in those directories.
In other words, those directories define the locations in which to create
package directories (which then must match the `package` delcaration in each
`.java` file.